requirements.txt
openai==1.3.5
requests
python-dotenv
python-telegram-bot==20.7
pytz